# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/accounts/admin/{tenant}/page/login": {
            "get": {
                "tags": [
                    "tenant-meta"
                ],
                "summary": "get tenant login page",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "租户",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "tenant-meta"
                ],
                "summary": "update tenant login page",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "租户名",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "bd",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/picture/{type}/upload": {
            "put": {
                "tags": [
                    "tenant-meta"
                ],
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "租户名",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "logo",
                        "description": "图片类型(background|logo)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/admin/{tenant}/proto": {
            "get": {
                "tags": [
                    "tenant-meta"
                ],
                "summary": "获取用户隐私协议",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "租户",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "tenant-meta"
                ],
                "summary": "update tenant login page",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "租户名",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "bd",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/accounts/{tenant}/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "login using username and password",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "login",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "next",
                        "name": "next",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/accounts/{tenant}/login/proto/{fileName}": {
            "get": {
                "tags": [
                    "login"
                ],
                "summary": "logout current user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "default",
                        "description": "fileName",
                        "name": "fileName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/accounts/{tenant}/logout": {
            "get": {
                "tags": [
                    "login"
                ],
                "summary": "logout current user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/accounts/{tenant}/register": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "register using username and password",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "tenant",
                        "name": "tenant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "login",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}